Implementação das foreign Keys

+++ src/main/java/com/funtec/systemhospital/repository/MedicoRepository.java
@@ -1,11 +1,9 @@
 package com.funtec.systemhospital.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.funtec.systemhospital.model.Medico;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicoRepository extends JpaRepository<Medico, Integer> {
-    
+public interface MedicoRepository extends JpaRepository<Medico, Long> {
 }
